AgentCPM-Explore/AgentDock/agentdock-node-explore/mcp_servers/read_file_enhanced/requirements.txt
camelot-py
markitdown[all]
mcp[cli]
toml
openai
pdfplumber